Refactor reverse to its own folder and separated files per type
