+++ test/unit/renderer/viz/expressions/reverse.test.js
@@ -1,4 +1,6 @@
 import { validateStaticType, validateMaxArgumentsError, validateCompileTypeError } from './utils';
+// import reverse from '../../../../../src/renderer/viz/expressions/reverse/reverse';
+// import { palettes } from '../../../../../src/renderer/viz/expressions';
 
 describe('src/renderer/viz/expressions/reverse', () => {
     describe('type', () => {
@@ -12,7 +14,15 @@ describe('src/renderer/viz/expressions/reverse', () => {
         validateMaxArgumentsError('reverse', [[1, 2], 0]);
         validateMaxArgumentsError('reverse', ['palette', 0]);
     });
-    describe('eval', () => {
-        // TODO check reversal of the array (see blend.test.js)
-    });
+    // describe('eval', () => {
+    //     // TODO check reversal of the array (see blend.test.js)
+    //     it('should reverse a palette', () => {
+    //         const maxColorIndex = [palettes.PRISM.length - 1];
+    //         let { 0: firstColor, maxColorIndex: lastColor } = palettes.PRISM;
+
+    //         const reversed = reverse(palettes.PRISM).eval();
+    //         expect(reversed[0]).toEqual(lastColor);
+    //         expect(reversed[maxColorIndex]).toEqual(firstColor);
+    //     });
+    // });
 });
